Add files via upload
After the user chooses a city that is populated, such as Boston, or logs into their organizational page, they are taken to a searchable map. The top address search bar allows the user to look at immigrant organizations closest to them. Once the user enters the starting address, a buffer zone appears around the location at the .5, 1, 3, and 5 mile radius. This allows the user to see if the client will need to walk or use transportation. This helps the user judge which organizations to put into the add to cart feature. The left side menu allows the user to search by service type. Once the user clicks on the service type, only the symbols correlated with that service type is shown on the maps symbology. Below each service type is a scrollable menu with all the organization names under that service type. Clicking on an individual organization will highlight that symbol in the map and pop-up.