Fix scrolling in the saved provider tab.
Also tweaks how the tabbed menu's size is determined:

- The overall size is set by the side-menu block, set to map height
- The side-menu block is a flex container. it's children are the menu
tabs and tab contents. The menu tabs use flex: initial to use their
native size. The tab panels (react-tabs__tab-panel blocks) use flex:
auto with a minimum height of 0 to shrink down to the available space in
the flex container.
- The tab contents blocks (service-providers and saved-list) are
contained in the tab panel blocks and use height: 100% to set their
heights to the available space.

Without this, the scrollable content is actually taller than the side
menu and the bottom of the scrollbar is cut off.