Fix drag-n-drop behavior.
We were overriding positioning of the draggable, which is not
recommended in the docs. This was causing an issue where the offset
between the mouse and draggable increased down the list, so dragging
around the end of the list would cause the draggables to go off-screen.
Removing the overrides makes the offset constant.